+++ src/client/sprites/3d/faces/octahedron.ts
@@ -1,0 +1,13 @@
+export const octahedronFaces: [number, number, ...number[]][] = [
+    [0, 2, 4],
+    [2, 1, 4],
+    [1, 3, 4],
+    [3, 0, 4],
+
+    [2, 0, 5],
+    [1, 2, 5],
+    [3, 1, 5],
+    [0, 3, 5]
+];
+
+export default octahedronFaces;