+++ src/client/sprites/3d/vertices/cube.ts
@@ -1,0 +1,16 @@
+import type { Vec3 } from '../../../types/Vectors.ts';
+
+const cubeVertices: Vec3[] = [
+    // Bottom
+    [-0.5, -0.5, -0.5],
+    [0.5, -0.5, -0.5],
+    [0.5, -0.5, 0.5],
+    [-0.5, -0.5, 0.5],
+    // Top
+    [-0.5, 0.5, -0.5],
+    [0.5, 0.5, -0.5],
+    [0.5, 0.5, 0.5],
+    [-0.5, 0.5, 0.5]
+];
+
+export default cubeVertices;