+++ src/client/sprites/3d/vertices/octahedron.ts
@@ -1,0 +1,12 @@
+import type { Vec3 } from '../../../types/Vectors.ts';
+
+export const octahedronVerticesVertices: Vec3[] = [
+    [ 1,  0,  0],
+    [-1,  0,  0],
+    [ 0,  1,  0],
+    [ 0, -1,  0],
+    [ 0,  0,  1],
+    [ 0,  0, -1]
+];
+
+export default octahedronVerticesVertices;