+++ src/client/sprites/3d/vertices/tetrahedron.ts
@@ -1,0 +1,10 @@
+import type { Vec3 } from '../../../types/Vectors.ts';
+
+export const tetrahedronVertices: Vec3[] = [
+    [ 1,  1,  1],
+    [-1, -1,  1],
+    [-1,  1, -1],
+    [ 1, -1, -1]
+];
+
+export default tetrahedronVertices;